Pharmacist Prescription DB
Added Pharmacist Prescription Functionality
- Can add medication and display properly

+++ MyApplication2/app/src/main/java/com/example/myapplication/Prescription.java
@@ -2,7 +2,7 @@ package com.example.myapplication;
 
 public class Prescription
 {
-    int drugId;
+    Long drugId;
     String drugPrescribed;
     Long dosage;
 
@@ -10,11 +10,11 @@ public class Prescription
 
     }
 
-    public int getDrugId() {
+    public Long getDrugId() {
         return drugId;
     }
 
-    public void setDrugId(int drugId) {
+    public void setDrugId(Long drugId) {
         this.drugId = drugId;
     }
